requirements.txt
requests
python-dotenv
tiktoken
openai
reportlab